tmp: Renamed Core/Singleton.mts to Core/Singletontmp.mts to avoid renaming issues for Windows users.
