+++ packages/architectura/src/Core/Singletontmp.mts
